merge

+++ InvalidPlayer/InvalidPlayer.Shared/Model/VideoItem.cs
@@ -8,7 +8,7 @@ namespace InvalidPlayer.Model
     {
         public string Url { get; set; }
 
-        public long Seconds { get; set; }
+        public double Seconds { get; set; }
 
         public int Size { get; set; }
     }
